PH-10: dry music platform buttons
